+++ lib/guard/test.rb
@@ -24,7 +24,7 @@ module Guard
     end
 
     def start
-      ::Guard::UI.info("Guard::Test #{TestVersion::VERSION} is running, with Test::Unit #{::Test::Unit::VERSION}!")
+      ::Guard::UI.info("Guard::Test #{TestVersion::VERSION} is running, with Test::Unit #{::Test::Unit::VERSION}!", :reset => true)
       run_all if @options[:all_on_start]
     end